pyproject.toml
[tool.poetry]
name = "currencycheck"
version = "0.1.0"
description = ""
authors = ["hubert <hubbertt.k@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
pytest = "^8.3.3"
yfinance = "^0.2.50"
pendulum = "^3.0.0"
vcrpy = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"